+++ tests/testthat/test-find.R
@@ -1,20 +1,32 @@
 context("test-find")
 
 
-# deprecated functions ----------------------------------------------------
+# moved functions ----------------------------------------------------
 
-test_that("can call to .Deprecated", {
+test_that(".Deprecated call contains function name", {
   f <- function() {
-    .Deprecated('x')
+    .Deprecated("pkg::x")
   }
 
-  expect_true(is_deprecated(f))
+  expect_false(has_moved("pkg", "foo", f))
+  expect_true(has_moved("pkg", "x", f))
 })
 
 test_that("returns FALSE for weird inputs", {
-  expect_false(is_deprecated(20))
-  expect_false(is_deprecated(mean))
+  expect_false(has_moved(obj = 20))
+  expect_false(has_moved(obj = mean))
 
   f <- function() {}
-  expect_false(is_deprecated(f))
+  expect_false(has_moved(obj = mean))
+
+  f <- function() {
+    .Deprecated()
+  }
+  expect_false(has_moved(obj = mean))
+
+  f <- function() {
+    .Deprecated(1)
+  }
+  expect_false(has_moved(obj = mean))
+
 })